Moving ConcurrentQueue to shared folder (#17956)
* Moving ConcurrentQueue to shared folder

Fixes: #17751
